dgram: do not emit error for failed DNS lookups in send()
Modifies the dgram send() method to not emit errors when a DNS lookup
fails if there is no callback. Given that the same UDP socket can be
used to send messages to different hosts, the socket can be reused even
if one of those send() fails.

This slightly changes the behavior of a stable API, so that it behaves
as users would expect to.

This is based of @chrisdickinson https://github.com/joyent/node/pull/7738.
Discussion in https://github.com/joyent/node/issues/4846.
