+++ lib/dgram.js
@@ -299,12 +299,8 @@ Socket.prototype.send = function(buffer,
 
   self._handle.lookup(address, function(ex, ip) {
     if (ex) {
-      if (callback) {
+      if (typeof callback === 'function') {
         callback(ex);
-
-        if (self.listeners('error').length)
-          self.emit('error', ex);
-
         return;
       }
